feat: Implement Customer Order History and Detail Screens
This commit introduces the following changes:

-   **Customer History Screen:**
    - Implemented `CustomerHistoryScreen` to display customer's order history.
    - Integrated with `CustomerViewModel` to fetch and display order data.
    - Implemented `OrderItem` composable to display individual order details.
    - Added logic to display order status with appropriate colors.
    - Added button to create new order.
-   -  Update `CustomerOrderScreen` to navigate to `CustomerHistoryScreen`.
-   **API Service:**
-   - Added `addAuthorizationHeader` function to `ApiService` to include bearer token in API requests.
-   **Customer Repository:**
-   - Modified `login` function in `CustomerRepositoryImpl` to add authorization header to API service after successful login.
-   - Change login state to manage state for each role.
-   **Model:**
-   - Change model to fit with api.
-   **Build Configuration:**
-   - Added dummy login credentials in `buildConfig` for testing.
-   **Other**
-    - Update all screen to fit with app.
+
+-   **Customer View Model:**
+    - Updated `CustomerState` sealed class to use data objects for Idle, Login, Register, ShopSearch, OrderHistory, and TrackOrder states.
+
+-   **API Service:**
+    - Modified the getOrderHistory function in ApiService to include the authorization token in the request headers.
+
+-   **Customer Repository:**
+    - Updated the login function in CustomerRepositoryImpl to set the token in the ApiService companion object upon successful login.
+
+-   **Order History Response:**
+    - Changed the type of totalPrice in OrderHistoryResponse from Int to Double.
+
+-   **Fake Module:**
+    - Updated the fakeOrderHistory function in FakeModule to return a list of OrderHistoryResponse objects with totalPrice as Double.

+++ app/src/main/java/com/example/manage/laundry/data/model/response/OrderHistoryResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class OrderHistoryResponse(
     val orderId: Int,
     val shopName: String,
-    val totalPrice: Int,
+    val totalPrice: Double,
     val status: String,
     val createdAt: String,
     val updatedAt: String
